ANDROID: KVM: arm64: Fix guest_share_host() with block-aligned range
In the very unlucky case where the guest requests sharing a PMD-aligned,
PMD-size range, on an existing table, where all entries are physically
contiguous, the mapping function will try to collapse the table into a
block.

This can happen with the following set of events:

  - Stage-2 page fault for addr X
  - Install a block mapping covering addr X
  - Guest relinquish addr X
     * Split the block containing addr X
     * Unmaps addr X
  - Guest share the entire block containing addr X
     * Install a mapping for addr X
     * kvm_pgtable_stage2_map() tries to collapse the table and run into
       the missing ->free_unlinked_table callback.

Huge-mappings must be only at request of the host, based on the
userspace stage-1 mapping and KVM memslot boundaries. Prevent any
potential table collapsing into block by rate-limiting the
sharing/unsharing range to half the size of a table.

While at it, add a free_unlinked_table() callback with a WARN to
simplify debug of spurious calls. This is safer than providing a
function callback that could silently break stage-2 annotations or
premissions.

Bug: 465276655
Bug: 419548963
Bug: 278011447
Change-Id: I4b94f38a2571dd9c5dab3fa41b4bb1133bec34a1
Signed-off-by: Vincent Donnefort <vdonnefort@google.com>